chore: apply cargo clippy --fix to remaining 6 warnings
Mechanical cleanups landing us at zero rustc + zero clippy warnings.

- mp31.rs: drop redundant `u32 -> u32` cast in `mp_montymul`.
- verify/helpers.rs: drop redundant `i16 -> i16` cast.
- q_derive.rs, sign/mod.rs: replace `x < lo || x >= hi` with
  `!Range::contains(&x)`.
- zint31.rs: replace `&tmp[..u].to_vec()` with `&tmp[..u]`. This one
  is a **real perf improvement** — it was allocating a fresh Vec per
  inner-loop iteration of `zint_rebuild_crt` during keygen. No
  functional change.
- q_derive.rs: keep `make_q001 -> Result<_, ()>` (sole caller uses Err
  purely as a retry flag with no payload worth surfacing); silenced
  clippy::result_unit_err with a targeted `#[allow]` + doc comment.

All 188 lib tests + all other suites remain green.

+++ src/keygen/mp31.rs
@@ -90,7 +90,7 @@ pub fn mp_half(a: u32, p: u32) -> u32 {
 #[inline]
 pub fn mp_montymul(a: u32, b: u32, p: u32, p0i: u32) -> u32 {
     let z = (a as u64).wrapping_mul(b as u64);
-    let w = ((z as u32).wrapping_mul(p0i)) as u32;
+    let w = (z as u32).wrapping_mul(p0i);
     let d = (z.wrapping_add((w as u64).wrapping_mul(p as u64)) >> 32) as u32;
     d.wrapping_sub(p)
         .wrapping_add(p & tbmask(d.wrapping_sub(p)))
